requirements.txt
Django==3.2.10
djangorestframework==3.12.4
django-cors-headers==3.10.0
django-filter==21.1
python-dotenv==0.19.2
gunicorn==20.1.0
psycopg2-binary==2.9.2